NO-ISSUE separate agent api from inventory module
