+++ internal/bminventory/agent_api.go
@@ -1,0 +1,393 @@
+package bminventory
+
+import (
+	"context"
+
+	"encoding/json"
+	"fmt"
+	"net/http"
+	"strings"
+	"time"
+
+	"github.com/go-openapi/runtime/middleware"
+	"github.com/go-openapi/strfmt"
+	"github.com/go-openapi/swag"
+	"github.com/jinzhu/gorm"
+	"github.com/openshift/assisted-service/internal/common"
+	"github.com/openshift/assisted-service/internal/host"
+	"github.com/openshift/assisted-service/internal/hostutil"
+	"github.com/openshift/assisted-service/internal/network"
+	"github.com/openshift/assisted-service/models"
+	"github.com/openshift/assisted-service/pkg/auth"
+	logutil "github.com/openshift/assisted-service/pkg/log"
+	"github.com/openshift/assisted-service/restapi/operations/installer"
+	"github.com/pkg/errors"
+	"github.com/sirupsen/logrus"
+)
+
+func (b *bareMetalInventory) RegisterHost(ctx context.Context, params installer.RegisterHostParams) middleware.Responder {
+	log := logutil.FromContext(ctx, b.log)
+	var host models.Host
+	var cluster common.Cluster
+	log.Infof("Register host: %+v", params)
+
+	if err := b.db.First(&cluster, "id = ?", params.ClusterID.String()).Error; err != nil {
+		log.WithError(err).Errorf("failed to get cluster: %s", params.ClusterID.String())
+		if gorm.IsRecordNotFoundError(err) {
+			return common.NewApiError(http.StatusNotFound, err)
+		}
+		return common.NewApiError(http.StatusInternalServerError, err)
+	}
+	err := b.db.First(&host, "id = ? and cluster_id = ?", *params.NewHostParams.HostID, params.ClusterID).Error
+	if err != nil && !gorm.IsRecordNotFoundError(err) {
+		log.WithError(err).Errorf("failed to get host %s in cluster: %s",
+			*params.NewHostParams.HostID, params.ClusterID.String())
+		return installer.NewRegisterHostInternalServerError().
+			WithPayload(common.GenerateError(http.StatusInternalServerError, err))
+	}
+
+	// In case host doesn't exists check if the cluster accept new hosts registration
+	if err != nil && gorm.IsRecordNotFoundError(err) {
+		if err = b.clusterApi.AcceptRegistration(&cluster); err != nil {
+			log.WithError(err).Errorf("failed to register host <%s> to cluster %s due to: %s",
+				params.NewHostParams.HostID, params.ClusterID.String(), err.Error())
+			b.eventsHandler.AddEvent(ctx, params.ClusterID, params.NewHostParams.HostID, models.EventSeverityError,
+				"Failed to register host: cluster cannot accept new hosts in its current state", time.Now())
+			return common.NewApiError(http.StatusConflict, err)
+		}
+	}
+
+	url := installer.GetHostURL{ClusterID: params.ClusterID, HostID: *params.NewHostParams.HostID}
+	kind := swag.String(models.HostKindHost)
+	switch swag.StringValue(cluster.Kind) {
+	case models.ClusterKindAddHostsCluster:
+		kind = swag.String(models.HostKindAddToExistingClusterHost)
+	case models.ClusterKindAddHostsOCPCluster:
+		kind = swag.String(models.HostKindAddToExistingClusterOCPHost)
+	}
+	host = models.Host{
+		ID:                    params.NewHostParams.HostID,
+		Href:                  swag.String(url.String()),
+		Kind:                  kind,
+		ClusterID:             params.ClusterID,
+		CheckedInAt:           strfmt.DateTime(time.Now()),
+		DiscoveryAgentVersion: params.NewHostParams.DiscoveryAgentVersion,
+		UserName:              auth.UserNameFromContext(ctx),
+		Role:                  models.HostRoleAutoAssign,
+	}
+
+	if err = b.hostApi.RegisterHost(ctx, &host); err != nil {
+		log.WithError(err).Errorf("failed to register host <%s> cluster <%s>",
+			params.NewHostParams.HostID.String(), params.ClusterID.String())
+		uerr := errors.Wrap(err, "Failed to register host: error creating host metadata")
+		b.eventsHandler.AddEvent(ctx, params.ClusterID, params.NewHostParams.HostID, models.EventSeverityError,
+			uerr.Error(), time.Now())
+		return returnRegisterHostTransitionError(http.StatusBadRequest, err)
+	}
+
+	if err = b.customizeHost(&host); err != nil {
+		b.eventsHandler.AddEvent(ctx, params.ClusterID, params.NewHostParams.HostID, models.EventSeverityError,
+			"Failed to register host: error setting host properties", time.Now())
+		return common.GenerateErrorResponder(err)
+	}
+
+	b.eventsHandler.AddEvent(ctx, params.ClusterID, params.NewHostParams.HostID, models.EventSeverityInfo,
+		fmt.Sprintf("Host %s: registered to cluster", hostutil.GetHostnameForMsg(&host)), time.Now())
+
+	hostRegistration := models.HostRegistrationResponse{
+		Host:                  host,
+		NextStepRunnerCommand: b.generateNextStepRunnerCommand(ctx, &params),
+	}
+
+	return installer.NewRegisterHostCreated().WithPayload(&hostRegistration)
+}
+
+func (b *bareMetalInventory) generateNextStepRunnerCommand(ctx context.Context, params *installer.RegisterHostParams) *models.HostRegistrationResponseAO1NextStepRunnerCommand {
+
+	currentImageTag := extractImageTag(b.AgentDockerImg)
+	if params.NewHostParams.DiscoveryAgentVersion != currentImageTag {
+		log := logutil.FromContext(ctx, b.log)
+		log.Infof("Host %s in cluster %s has outdated agent image %s, updating to %s",
+			params.NewHostParams.HostID.String(), params.ClusterID.String(), params.NewHostParams.DiscoveryAgentVersion, currentImageTag)
+	}
+
+	config := host.NextStepRunnerConfig{
+		ServiceBaseURL:       b.ServiceBaseURL,
+		ClusterID:            params.ClusterID.String(),
+		HostID:               params.NewHostParams.HostID.String(),
+		UseCustomCACert:      b.ServiceCACertPath != "",
+		NextStepRunnerImage:  b.AgentDockerImg,
+		SkipCertVerification: b.SkipCertVerification,
+	}
+	command, args := host.GetNextStepRunnerCommand(&config)
+	return &models.HostRegistrationResponseAO1NextStepRunnerCommand{
+		Command: command,
+		Args:    *args,
+	}
+}
+
+func extractImageTag(fullName string) string {
+	suffix := strings.Split(fullName, ":")
+	return suffix[len(suffix)-1]
+}
+
+func returnRegisterHostTransitionError(
+	defaultCode int32,
+	err error) middleware.Responder {
+	if isRegisterHostForbiddenDueWrongBootOrder(err) {
+		return installer.NewRegisterHostForbidden().WithPayload(
+			&models.InfraError{
+				Code:    swag.Int32(http.StatusForbidden),
+				Message: swag.String(err.Error()),
+			})
+	}
+	return common.NewApiError(defaultCode, err)
+}
+
+func isRegisterHostForbiddenDueWrongBootOrder(err error) bool {
+	if serr, ok := err.(*common.ApiErrorResponse); ok {
+		return serr.StatusCode() == http.StatusForbidden
+	}
+	return false
+}
+
+func (b *bareMetalInventory) GetNextSteps(ctx context.Context, params installer.GetNextStepsParams) middleware.Responder {
+	log := logutil.FromContext(ctx, b.log)
+	var steps models.Steps
+	var host models.Host
+
+	txSuccess := false
+	tx := b.db.Begin()
+	defer func() {
+		if !txSuccess {
+			log.Error("get next steps failed")
+			tx.Rollback()
+		}
+		if r := recover(); r != nil {
+			log.Error("get next steps failed")
+			tx.Rollback()
+		}
+	}()
+
+	if tx.Error != nil {
+		log.WithError(tx.Error).Errorf("failed to start db transaction")
+		return installer.NewUpdateClusterInternalServerError().
+			WithPayload(common.GenerateError(http.StatusInternalServerError, errors.New("DB error, failed to start transaction")))
+	}
+
+	//TODO check the error type
+	if err := tx.First(&host, "id = ? and cluster_id = ?", params.HostID, params.ClusterID).Error; err != nil {
+		log.WithError(err).Errorf("failed to find host: %s", params.HostID)
+		return installer.NewGetNextStepsNotFound().
+			WithPayload(common.GenerateError(http.StatusNotFound, err))
+	}
+
+	host.CheckedInAt = strfmt.DateTime(time.Now())
+	if err := tx.Model(&host).Update("checked_in_at", host.CheckedInAt).Error; err != nil {
+		log.WithError(err).Errorf("failed to update host: %s", params.ClusterID)
+		return installer.NewGetNextStepsInternalServerError()
+	}
+
+	if err := tx.Commit().Error; err != nil {
+		log.Error(err)
+		return installer.NewGetNextStepsInternalServerError()
+	}
+	txSuccess = true
+
+	var err error
+	steps, err = b.hostApi.GetNextSteps(ctx, &host)
+	if err != nil {
+		log.WithError(err).Errorf("failed to get steps for host %s cluster %s", params.HostID, params.ClusterID)
+	}
+
+	return installer.NewGetNextStepsOK().WithPayload(&steps)
+}
+
+func (b *bareMetalInventory) PostStepReply(ctx context.Context, params installer.PostStepReplyParams) middleware.Responder {
+	var err error
+	log := logutil.FromContext(ctx, b.log)
+	msg := fmt.Sprintf("Received step reply <%s> from cluster <%s> host <%s>  exit-code <%d> stdout <%s> stderr <%s>", params.Reply.StepID, params.ClusterID,
+		params.HostID, params.Reply.ExitCode, params.Reply.Output, params.Reply.Error)
+
+	var host models.Host
+	if err = b.db.First(&host, "id = ? and cluster_id = ?", params.HostID, params.ClusterID).Error; err != nil {
+		log.WithError(err).Errorf("Failed to find host <%s> cluster <%s> step <%s> exit code %d stdout <%s> stderr <%s>",
+			params.HostID, params.ClusterID, params.Reply.StepID, params.Reply.ExitCode, params.Reply.Output, params.Reply.Error)
+		return installer.NewPostStepReplyNotFound().
+			WithPayload(common.GenerateError(http.StatusNotFound, err))
+	}
+
+	//check the output exit code
+	if params.Reply.ExitCode != 0 {
+		err = errors.New(msg)
+		log.WithError(err).Errorf("Exit code is <%d> ", params.Reply.ExitCode)
+		handlingError := b.handleReplyError(params, ctx, log, &host)
+		if handlingError != nil {
+			log.WithError(handlingError).Errorf("Failed handling reply error for host <%s> cluster <%s>", params.HostID, params.ClusterID)
+		}
+		return installer.NewPostStepReplyNoContent()
+	}
+
+	log.Infof(msg)
+
+	var stepReply string
+	stepReply, err = filterReplyByType(params)
+	if err != nil {
+		log.WithError(err).Errorf("Failed decode <%s> reply for host <%s> cluster <%s>",
+			params.Reply.StepID, params.HostID, params.ClusterID)
+		return installer.NewPostStepReplyBadRequest().
+			WithPayload(common.GenerateError(http.StatusBadRequest, err))
+	}
+
+	err = handleReplyByType(params, b, ctx, host, stepReply)
+	if err != nil {
+		log.WithError(err).Errorf("Failed to update step reply for host <%s> cluster <%s> step <%s>",
+			params.HostID, params.ClusterID, params.Reply.StepID)
+		return installer.NewPostStepReplyInternalServerError().
+			WithPayload(common.GenerateError(http.StatusInternalServerError, err))
+	}
+
+	return installer.NewPostStepReplyNoContent()
+}
+
+func (b *bareMetalInventory) handleReplyError(params installer.PostStepReplyParams, ctx context.Context, log logrus.FieldLogger, h *models.Host) error {
+
+	if params.Reply.StepType == models.StepTypeInstall {
+		// Handle case of installation error due to an already running assisted-installer.
+		if params.Reply.ExitCode == ContainerAlreadyRunningExitCode && strings.Contains(params.Reply.Error, "the container name \"assisted-installer\" is already in use") {
+			log.Warnf("Install command failed due to an already running installation: %s", params.Reply.Error)
+			return nil
+		}
+		//if it's install step - need to move host to error
+		return b.hostApi.HandleInstallationFailure(ctx, h)
+	}
+	return nil
+}
+
+func (b *bareMetalInventory) updateFreeAddressesReport(ctx context.Context, host *models.Host, freeAddressesReport string) error {
+	var (
+		err           error
+		freeAddresses models.FreeNetworksAddresses
+	)
+	log := logutil.FromContext(ctx, b.log)
+	if err = json.Unmarshal([]byte(freeAddressesReport), &freeAddresses); err != nil {
+		log.WithError(err).Warnf("Json unmarshal free addresses of host %s", host.ID.String())
+		return err
+	}
+	if len(freeAddresses) == 0 {
+		err = errors.Errorf("Free addresses for host %s is empty", host.ID.String())
+		log.WithError(err).Warn("Update free addresses")
+		return err
+	}
+	if err = b.db.Model(&models.Host{}).Where("id = ? and cluster_id = ?", host.ID.String(),
+		host.ClusterID.String()).Updates(map[string]interface{}{"free_addresses": freeAddressesReport}).Error; err != nil {
+		log.WithError(err).Warnf("Update free addresses of host %s", host.ID.String())
+		return err
+	}
+	// Gorm sets the number of changed rows in AffectedRows and not the number of matched rows.  Therefore, if the report hasn't changed
+	// from the previous report, the AffectedRows will be 0 but it will still be correct.  So no error reporting needed for AffectedRows == 0
+	return nil
+}
+
+func (b *bareMetalInventory) processDhcpAllocationResponse(ctx context.Context, host *models.Host, dhcpAllocationResponseStr string) error {
+	var (
+		err                   error
+		dhcpAllocationReponse models.DhcpAllocationResponse
+		cluster               common.Cluster
+	)
+	log := logutil.FromContext(ctx, b.log)
+	if err = b.db.Take(&cluster, "id = ?", host.ClusterID.String()).Error; err != nil {
+		log.WithError(err).Warnf("Get cluster %s", host.ClusterID.String())
+		return err
+	}
+	if !swag.BoolValue(cluster.VipDhcpAllocation) {
+		err = errors.Errorf("DHCP not enabled in cluster %s", host.ClusterID.String())
+		log.WithError(err).Warn("processDhcpAllocationResponse")
+		return err
+	}
+	if err = json.Unmarshal([]byte(dhcpAllocationResponseStr), &dhcpAllocationReponse); err != nil {
+		log.WithError(err).Warnf("Json unmarshal dhcp allocation from host %s", host.ID.String())
+		return err
+	}
+	apiVip := dhcpAllocationReponse.APIVipAddress.String()
+	ingressVip := dhcpAllocationReponse.IngressVipAddress.String()
+	isApiVipInMachineCIDR, err := network.IpInCidr(apiVip, cluster.MachineNetworkCidr)
+	if err != nil {
+		log.WithError(err).Warn("Ip in CIDR for API VIP")
+		return err
+	}
+
+	isIngressVipInMachineCIDR, err := network.IpInCidr(ingressVip, cluster.MachineNetworkCidr)
+	if err != nil {
+		log.WithError(err).Warn("Ip in CIDR for Ingress VIP")
+		return err
+	}
+
+	if !(isApiVipInMachineCIDR && isIngressVipInMachineCIDR) {
+		err = errors.Errorf("At least of the IPs (%s, %s) is not in machine CIDR %s", apiVip, ingressVip, cluster.MachineNetworkCidr)
+		log.WithError(err).Warn("IP in CIDR")
+		return err
+	}
+
+	err = network.VerifyLease(dhcpAllocationReponse.APIVipLease)
+	if err != nil {
+		log.WithError(err).Warnf("API Vip not validated")
+		return err
+	}
+	err = network.VerifyLease(dhcpAllocationReponse.IngressVipLease)
+	if err != nil {
+		log.WithError(err).Warnf("Ingress Vip not validated")
+		return err
+	}
+	return b.clusterApi.SetVipsData(ctx, &cluster, apiVip, ingressVip, dhcpAllocationReponse.APIVipLease, dhcpAllocationReponse.IngressVipLease, b.db)
+}
+
+func handleReplyByType(params installer.PostStepReplyParams, b *bareMetalInventory, ctx context.Context, host models.Host, stepReply string) error {
+	var err error
+	switch params.Reply.StepType {
+	case models.StepTypeInventory:
+		err = b.hostApi.UpdateInventory(ctx, &host, stepReply)
+	case models.StepTypeConnectivityCheck:
+		err = b.hostApi.UpdateConnectivityReport(ctx, &host, stepReply)
+	case models.StepTypeAPIVipConnectivityCheck:
+		err = b.hostApi.UpdateApiVipConnectivityReport(ctx, &host, stepReply)
+	case models.StepTypeFreeNetworkAddresses:
+		err = b.updateFreeAddressesReport(ctx, &host, stepReply)
+	case models.StepTypeDhcpLeaseAllocate:
+		err = b.processDhcpAllocationResponse(ctx, &host, stepReply)
+	}
+	return err
+}
+
+func filterReplyByType(params installer.PostStepReplyParams) (string, error) {
+	var stepReply string
+	var err error
+
+	// To make sure we store only information defined in swagger we unmarshal and marshal the stepReplyParams.
+	switch params.Reply.StepType {
+	case models.StepTypeInventory:
+		stepReply, err = filterReply(&models.Inventory{}, params.Reply.Output)
+	case models.StepTypeConnectivityCheck:
+		stepReply, err = filterReply(&models.ConnectivityReport{}, params.Reply.Output)
+	case models.StepTypeAPIVipConnectivityCheck:
+		stepReply, err = filterReply(&models.APIVipConnectivityResponse{}, params.Reply.Output)
+	case models.StepTypeFreeNetworkAddresses:
+		stepReply, err = filterReply(&models.FreeNetworksAddresses{}, params.Reply.Output)
+	case models.StepTypeDhcpLeaseAllocate:
+		stepReply, err = filterReply(&models.DhcpAllocationResponse{}, params.Reply.Output)
+	}
+
+	return stepReply, err
+}
+
+// filterReply return only the expected parameters from the input.
+func filterReply(expected interface{}, input string) (string, error) {
+	if err := json.Unmarshal([]byte(input), expected); err != nil {
+		return "", err
+	}
+	reply, err := json.Marshal(expected)
+	if err != nil {
+		return "", err
+	}
+	return string(reply), nil
+}
