+++ internal/bminventory/inventory_test.go
@@ -460,618 +460,6 @@ func createCluster(db *gorm.DB, status string) *common.Cluster {
 	return c
 }
 
-var _ = Describe("RegisterHost", func() {
-	var (
-		bm                  *bareMetalInventory
-		cfg                 Config
-		db                  *gorm.DB
-		ctx                 = context.Background()
-		dbName              = "register_host_api"
-		hostID              strfmt.UUID
-		ctrl                *gomock.Controller
-		mockClusterAPI      *cluster.MockAPI
-		mockHostAPI         *host.MockAPI
-		mockEventsHandler   *events.MockHandler
-		mockSecretValidator *validations.MockPullSecretValidator
-	)
-
-	BeforeEach(func() {
-		ctrl = gomock.NewController(GinkgoT())
-		mockClusterAPI = cluster.NewMockAPI(ctrl)
-		mockHostAPI = host.NewMockAPI(ctrl)
-		mockEventsHandler = events.NewMockHandler(ctrl)
-		mockSecretValidator = validations.NewMockPullSecretValidator(ctrl)
-		hostID = strfmt.UUID(uuid.New().String())
-		db = common.PrepareTestDB(dbName)
-		bm = NewBareMetalInventory(db, getTestLog(), mockHostAPI, mockClusterAPI, cfg, nil, mockEventsHandler,
-			nil, nil, getTestAuthHandler(), nil, nil, mockSecretValidator)
-	})
-
-	AfterEach(func() {
-		common.DeleteTestDB(db, dbName)
-		ctrl.Finish()
-	})
-
-	It("register host to non-existing cluster", func() {
-		reply := bm.RegisterHost(ctx, installer.RegisterHostParams{
-			ClusterID: strfmt.UUID(uuid.New().String()),
-			NewHostParams: &models.HostCreateParams{
-				DiscoveryAgentVersion: "v1",
-				HostID:                &hostID,
-			},
-		})
-		apiErr, ok := reply.(*common.ApiErrorResponse)
-		Expect(ok).Should(BeTrue())
-		Expect(apiErr.StatusCode()).Should(Equal(int32(http.StatusNotFound)))
-	})
-
-	It("register host to a cluster while installation is in progress", func() {
-		By("creating the cluster")
-		cluster := createCluster(db, models.ClusterStatusInstalling)
-
-		allowedStates := []string{
-			models.ClusterStatusInsufficient, models.ClusterStatusReady,
-			models.ClusterStatusPendingForInput, models.ClusterStatusAddingHosts}
-		err := errors.Errorf(
-			"Cluster %s is in installing state, host can register only in one of %s",
-			cluster.ID, allowedStates)
-
-		mockClusterAPI.EXPECT().AcceptRegistration(gomock.Any()).Return(err).Times(1)
-
-		mockEventsHandler.EXPECT().
-			AddEvent(gomock.Any(), *cluster.ID, &hostID, models.EventSeverityError, gomock.Any(), gomock.Any()).
-			Times(1)
-
-		By("trying to register an host while installation takes place")
-		reply := bm.RegisterHost(ctx, installer.RegisterHostParams{
-			ClusterID: *cluster.ID,
-			NewHostParams: &models.HostCreateParams{
-				DiscoveryAgentVersion: "v1",
-				HostID:                &hostID,
-			},
-		})
-
-		By("verifying returned response")
-		apiErr, ok := reply.(*common.ApiErrorResponse)
-		Expect(ok).Should(BeTrue())
-		Expect(apiErr.StatusCode()).Should(Equal(int32(http.StatusConflict)))
-		Expect(apiErr.Error()).Should(Equal(err.Error()))
-	})
-
-	It("register_success", func() {
-		cluster := createCluster(db, models.ClusterStatusInsufficient)
-
-		mockClusterAPI.EXPECT().AcceptRegistration(gomock.Any()).Return(nil).Times(1)
-		mockHostAPI.EXPECT().RegisterHost(gomock.Any(), gomock.Any()).
-			DoAndReturn(func(ctx context.Context, h *models.Host) error {
-				// validate that host is registered with auto-assign role
-				Expect(h.Role).Should(Equal(models.HostRoleAutoAssign))
-				return nil
-			}).Times(1)
-		mockHostAPI.EXPECT().GetStagesByRole(gomock.Any(), gomock.Any()).Return(nil).Times(1)
-		mockEventsHandler.EXPECT().
-			AddEvent(gomock.Any(), *cluster.ID, &hostID, models.EventSeverityInfo, gomock.Any(), gomock.Any()).
-			Times(1)
-
-		bm.ServiceBaseURL = uuid.New().String()
-		bm.ServiceCACertPath = uuid.New().String()
-		bm.AgentDockerImg = uuid.New().String()
-		bm.SkipCertVerification = true
-
-		reply := bm.RegisterHost(ctx, installer.RegisterHostParams{
-			ClusterID: *cluster.ID,
-			NewHostParams: &models.HostCreateParams{
-				DiscoveryAgentVersion: "v1",
-				HostID:                &hostID,
-			},
-		})
-		_, ok := reply.(*installer.RegisterHostCreated)
-		Expect(ok).Should(BeTrue())
-
-		By("register_returns_next_step_runner_command")
-		payload := reply.(*installer.RegisterHostCreated).Payload
-		Expect(payload).ShouldNot(BeNil())
-		command := payload.NextStepRunnerCommand
-		Expect(command).ShouldNot(BeNil())
-		Expect(command.Command).ShouldNot(BeEmpty())
-		Expect(command.Args).ShouldNot(BeEmpty())
-	})
-
-	It("host_api_failure", func() {
-		cluster := createCluster(db, models.ClusterStatusInsufficient)
-		expectedErrMsg := "some-internal-error"
-
-		mockClusterAPI.EXPECT().AcceptRegistration(gomock.Any()).Return(nil).Times(1)
-		mockHostAPI.EXPECT().RegisterHost(gomock.Any(), gomock.Any()).Return(errors.New(expectedErrMsg)).Times(1)
-		mockEventsHandler.EXPECT().
-			AddEvent(gomock.Any(), *cluster.ID, &hostID, models.EventSeverityError, gomock.Any(), gomock.Any()).
-			Times(1)
-		reply := bm.RegisterHost(ctx, installer.RegisterHostParams{
-			ClusterID: *cluster.ID,
-			NewHostParams: &models.HostCreateParams{
-				DiscoveryAgentVersion: "v1",
-				HostID:                &hostID,
-			},
-		})
-		err, ok := reply.(*common.ApiErrorResponse)
-		Expect(ok).Should(BeTrue())
-		Expect(err.StatusCode()).Should(Equal(int32(http.StatusBadRequest)))
-		Expect(err.Error()).Should(ContainSubstring(expectedErrMsg))
-	})
-})
-
-var _ = Describe("GetNextSteps", func() {
-	var (
-		bm                  *bareMetalInventory
-		cfg                 Config
-		db                  *gorm.DB
-		ctx                 = context.Background()
-		ctrl                *gomock.Controller
-		mockHostApi         *host.MockAPI
-		mockEvents          *events.MockHandler
-		mockSecretValidator *validations.MockPullSecretValidator
-		defaultNextStepIn   int64
-		dbName              = "get_next_steps"
-	)
-
-	BeforeEach(func() {
-		Expect(envconfig.Process("test", &cfg)).ShouldNot(HaveOccurred())
-		ctrl = gomock.NewController(GinkgoT())
-		defaultNextStepIn = 60
-		db = common.PrepareTestDB(dbName)
-		mockHostApi = host.NewMockAPI(ctrl)
-		mockEvents = events.NewMockHandler(ctrl)
-		mockSecretValidator = validations.NewMockPullSecretValidator(ctrl)
-		bm = NewBareMetalInventory(db, getTestLog(), mockHostApi, nil, cfg, nil, mockEvents, nil, nil, getTestAuthHandler(), nil, nil, mockSecretValidator)
-	})
-
-	AfterEach(func() {
-		ctrl.Finish()
-		common.DeleteTestDB(db, dbName)
-	})
-
-	It("get_next_steps_unknown_host", func() {
-		clusterId := strToUUID(uuid.New().String())
-		unregistered_hostID := strToUUID(uuid.New().String())
-
-		generateReply := bm.GetNextSteps(ctx, installer.GetNextStepsParams{
-			ClusterID: *clusterId,
-			HostID:    *unregistered_hostID,
-		})
-		Expect(generateReply).Should(BeAssignableToTypeOf(installer.NewGetNextStepsNotFound()))
-	})
-
-	It("get_next_steps_success", func() {
-		clusterId := strToUUID(uuid.New().String())
-		hostId := strToUUID(uuid.New().String())
-		host := models.Host{
-			ID:        hostId,
-			ClusterID: *clusterId,
-			Status:    swag.String("discovering"),
-		}
-		Expect(db.Create(&host).Error).ShouldNot(HaveOccurred())
-
-		var err error
-		expectedStepsReply := models.Steps{NextInstructionSeconds: defaultNextStepIn, Instructions: []*models.Step{{StepType: models.StepTypeInventory},
-			{StepType: models.StepTypeConnectivityCheck}}}
-		mockHostApi.EXPECT().GetNextSteps(gomock.Any(), gomock.Any()).Return(expectedStepsReply, err)
-		reply := bm.GetNextSteps(ctx, installer.GetNextStepsParams{
-			ClusterID: *clusterId,
-			HostID:    *hostId,
-		})
-		Expect(reply).Should(BeAssignableToTypeOf(installer.NewGetNextStepsOK()))
-		stepsReply := reply.(*installer.GetNextStepsOK).Payload
-		expectedStepsType := []models.StepType{models.StepTypeInventory, models.StepTypeConnectivityCheck}
-		Expect(stepsReply.Instructions).To(HaveLen(len(expectedStepsType)))
-		for i, step := range stepsReply.Instructions {
-			Expect(step.StepType).Should(Equal(expectedStepsType[i]))
-		}
-	})
-})
-
-func makeFreeAddresses(network string, ips ...strfmt.IPv4) *models.FreeNetworkAddresses {
-	return &models.FreeNetworkAddresses{
-		FreeAddresses: ips,
-		Network:       network,
-	}
-}
-
-func makeFreeNetworksAddresses(elems ...*models.FreeNetworkAddresses) models.FreeNetworksAddresses {
-	return models.FreeNetworksAddresses(elems)
-}
-
-func makeFreeNetworksAddressesStr(elems ...*models.FreeNetworkAddresses) string {
-	toMarshal := models.FreeNetworksAddresses(elems)
-	b, err := json.Marshal(&toMarshal)
-	Expect(err).ToNot(HaveOccurred())
-	return string(b)
-}
-
-var _ = Describe("PostStepReply", func() {
-	var (
-		bm                  *bareMetalInventory
-		cfg                 Config
-		db                  *gorm.DB
-		ctx                 = context.Background()
-		ctrl                *gomock.Controller
-		mockClusterApi      *cluster.MockAPI
-		mockHostApi         *host.MockAPI
-		mockEvents          *events.MockHandler
-		mockSecretValidator *validations.MockPullSecretValidator
-		dbName              = "post_step_reply"
-	)
-
-	BeforeEach(func() {
-		Expect(envconfig.Process("test", &cfg)).ShouldNot(HaveOccurred())
-		ctrl = gomock.NewController(GinkgoT())
-		db = common.PrepareTestDB(dbName)
-		mockHostApi = host.NewMockAPI(ctrl)
-		mockEvents = events.NewMockHandler(ctrl)
-		mockClusterApi = cluster.NewMockAPI(ctrl)
-		mockSecretValidator = validations.NewMockPullSecretValidator(ctrl)
-		bm = NewBareMetalInventory(db, getTestLog(), mockHostApi, mockClusterApi, cfg, nil, mockEvents, nil, nil, getTestAuthHandler(), nil, nil, mockSecretValidator)
-	})
-
-	AfterEach(func() {
-		ctrl.Finish()
-		common.DeleteTestDB(db, dbName)
-	})
-
-	Context("Free addresses", func() {
-		var makeStepReply = func(clusterID, hostID strfmt.UUID, freeAddresses models.FreeNetworksAddresses) installer.PostStepReplyParams {
-			b, _ := json.Marshal(&freeAddresses)
-			return installer.PostStepReplyParams{
-				ClusterID: clusterID,
-				HostID:    hostID,
-				Reply: &models.StepReply{
-					Output:   string(b),
-					StepType: models.StepTypeFreeNetworkAddresses,
-				},
-			}
-		}
-
-		It("free addresses success", func() {
-			clusterId := strToUUID(uuid.New().String())
-			hostId := strToUUID(uuid.New().String())
-			host := models.Host{
-				ID:        hostId,
-				ClusterID: *clusterId,
-				Status:    swag.String("discovering"),
-			}
-			Expect(db.Create(&host).Error).ShouldNot(HaveOccurred())
-			toMarshal := makeFreeNetworksAddresses(makeFreeAddresses("10.0.0.0/24", "10.0.0.0", "10.0.0.1"))
-			params := makeStepReply(*clusterId, *hostId, toMarshal)
-			reply := bm.PostStepReply(ctx, params)
-			Expect(reply).Should(BeAssignableToTypeOf(installer.NewPostStepReplyNoContent()))
-			var h models.Host
-			Expect(db.Take(&h, "cluster_id = ? and id = ?", clusterId.String(), hostId.String()).Error).ToNot(HaveOccurred())
-			var f models.FreeNetworksAddresses
-			Expect(json.Unmarshal([]byte(h.FreeAddresses), &f)).ToNot(HaveOccurred())
-			Expect(&f).To(Equal(&toMarshal))
-		})
-
-		It("free addresses empty", func() {
-			clusterId := strToUUID(uuid.New().String())
-			hostId := strToUUID(uuid.New().String())
-			host := models.Host{
-				ID:        hostId,
-				ClusterID: *clusterId,
-				Status:    swag.String("discovering"),
-			}
-			Expect(db.Create(&host).Error).ShouldNot(HaveOccurred())
-			toMarshal := makeFreeNetworksAddresses()
-			params := makeStepReply(*clusterId, *hostId, toMarshal)
-			reply := bm.PostStepReply(ctx, params)
-			Expect(reply).Should(BeAssignableToTypeOf(installer.NewPostStepReplyInternalServerError()))
-			var h models.Host
-			Expect(db.Take(&h, "cluster_id = ? and id = ?", clusterId.String(), hostId.String()).Error).ToNot(HaveOccurred())
-			Expect(h.FreeAddresses).To(BeEmpty())
-		})
-	})
-
-	Context("Dhcp allocation", func() {
-		var (
-			clusterId, hostId *strfmt.UUID
-			makeStepReply     = func(clusterID, hostID strfmt.UUID, dhcpAllocationResponse *models.DhcpAllocationResponse) installer.PostStepReplyParams {
-				b, err := json.Marshal(dhcpAllocationResponse)
-				Expect(err).ToNot(HaveOccurred())
-				return installer.PostStepReplyParams{
-					ClusterID: clusterID,
-					HostID:    hostID,
-					Reply: &models.StepReply{
-						Output:   string(b),
-						StepType: models.StepTypeDhcpLeaseAllocate,
-					},
-				}
-			}
-			makeResponse = func(apiVipStr, ingressVipStr string) *models.DhcpAllocationResponse {
-				apiVip := strfmt.IPv4(apiVipStr)
-				ingressVip := strfmt.IPv4(ingressVipStr)
-				ret := models.DhcpAllocationResponse{
-					APIVipAddress:     &apiVip,
-					IngressVipAddress: &ingressVip,
-				}
-				return &ret
-			}
-			makeResponseWithLeases = func(apiVipStr, ingressVipStr, apiLease, ingressLease string) *models.DhcpAllocationResponse {
-				apiVip := strfmt.IPv4(apiVipStr)
-				ingressVip := strfmt.IPv4(ingressVipStr)
-				ret := models.DhcpAllocationResponse{
-					APIVipAddress:     &apiVip,
-					IngressVipAddress: &ingressVip,
-					APIVipLease:       apiLease,
-					IngressVipLease:   ingressLease,
-				}
-				return &ret
-			}
-		)
-		BeforeEach(func() {
-			clusterId = strToUUID(uuid.New().String())
-			hostId = strToUUID(uuid.New().String())
-			host := models.Host{
-				ID:        hostId,
-				ClusterID: *clusterId,
-				Status:    swag.String("insufficient"),
-			}
-			Expect(db.Create(&host).Error).ShouldNot(HaveOccurred())
-		})
-		It("Happy flow with leases", func() {
-			cluster := common.Cluster{
-				Cluster: models.Cluster{
-					ID:                 clusterId,
-					VipDhcpAllocation:  swag.Bool(true),
-					MachineNetworkCidr: "1.2.3.0/24",
-					Status:             swag.String(models.ClusterStatusInsufficient),
-				},
-			}
-			Expect(db.Create(&cluster).Error).ToNot(HaveOccurred())
-			params := makeStepReply(*clusterId, *hostId, makeResponseWithLeases("1.2.3.10", "1.2.3.11", "lease { hello abc; }", "lease { hello abc; }"))
-			mockClusterApi.EXPECT().SetVipsData(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Return(nil)
-			reply := bm.PostStepReply(ctx, params)
-			Expect(reply).Should(BeAssignableToTypeOf(installer.NewPostStepReplyNoContent()))
-		})
-		It("API lease invalid", func() {
-			cluster := common.Cluster{
-				Cluster: models.Cluster{
-					ID:                 clusterId,
-					VipDhcpAllocation:  swag.Bool(true),
-					MachineNetworkCidr: "1.2.3.0/24",
-					Status:             swag.String(models.ClusterStatusInsufficient),
-				},
-			}
-			Expect(db.Create(&cluster).Error).ToNot(HaveOccurred())
-			params := makeStepReply(*clusterId, *hostId, makeResponseWithLeases("1.2.3.10", "1.2.3.11", "llease { hello abc; }", "lease { hello abc; }"))
-			reply := bm.PostStepReply(ctx, params)
-			Expect(reply).Should(BeAssignableToTypeOf(installer.NewPostStepReplyInternalServerError()))
-		})
-		It("Happy flow", func() {
-			cluster := common.Cluster{
-				Cluster: models.Cluster{
-					ID:                 clusterId,
-					VipDhcpAllocation:  swag.Bool(true),
-					MachineNetworkCidr: "1.2.3.0/24",
-					Status:             swag.String(models.ClusterStatusInsufficient),
-				},
-			}
-			Expect(db.Create(&cluster).Error).ToNot(HaveOccurred())
-			params := makeStepReply(*clusterId, *hostId, makeResponse("1.2.3.10", "1.2.3.11"))
-			mockClusterApi.EXPECT().SetVipsData(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Return(nil)
-			reply := bm.PostStepReply(ctx, params)
-			Expect(reply).Should(BeAssignableToTypeOf(installer.NewPostStepReplyNoContent()))
-		})
-		It("DHCP not enabled", func() {
-			cluster := common.Cluster{
-				Cluster: models.Cluster{
-					ID:                 clusterId,
-					VipDhcpAllocation:  swag.Bool(false),
-					MachineNetworkCidr: "1.2.3.0/24",
-					Status:             swag.String(models.ClusterStatusInsufficient),
-				},
-			}
-			Expect(db.Create(&cluster).Error).ToNot(HaveOccurred())
-			params := makeStepReply(*clusterId, *hostId, makeResponse("1.2.3.10", "1.2.3.11"))
-			reply := bm.PostStepReply(ctx, params)
-			Expect(reply).Should(BeAssignableToTypeOf(installer.NewPostStepReplyInternalServerError()))
-		})
-		It("Bad ingress VIP", func() {
-			cluster := common.Cluster{
-				Cluster: models.Cluster{
-					ID:                 clusterId,
-					VipDhcpAllocation:  swag.Bool(true),
-					MachineNetworkCidr: "1.2.3.0/24",
-					Status:             swag.String(models.ClusterStatusInsufficient),
-				},
-			}
-			Expect(db.Create(&cluster).Error).ToNot(HaveOccurred())
-			params := makeStepReply(*clusterId, *hostId, makeResponse("1.2.3.10", "1.2.4.11"))
-			reply := bm.PostStepReply(ctx, params)
-			Expect(reply).Should(BeAssignableToTypeOf(installer.NewPostStepReplyInternalServerError()))
-		})
-		It("New IPs while in insufficient", func() {
-			cluster := common.Cluster{
-				Cluster: models.Cluster{
-					ID:                 clusterId,
-					VipDhcpAllocation:  swag.Bool(true),
-					MachineNetworkCidr: "1.2.3.0/24",
-					APIVip:             "1.2.3.20",
-					IngressVip:         "1.2.3.11",
-					Status:             swag.String(models.ClusterStatusInsufficient),
-				},
-			}
-			Expect(db.Create(&cluster).Error).ToNot(HaveOccurred())
-			params := makeStepReply(*clusterId, *hostId, makeResponse("1.2.3.10", "1.2.3.11"))
-			mockClusterApi.EXPECT().SetVipsData(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Return(nil)
-			reply := bm.PostStepReply(ctx, params)
-			Expect(reply).Should(BeAssignableToTypeOf(installer.NewPostStepReplyNoContent()))
-		})
-		It("New IPs while in installing", func() {
-			cluster := common.Cluster{
-				Cluster: models.Cluster{
-					ID:                 clusterId,
-					VipDhcpAllocation:  swag.Bool(true),
-					MachineNetworkCidr: "1.2.3.0/24",
-					APIVip:             "1.2.3.20",
-					IngressVip:         "1.2.3.11",
-					Status:             swag.String(models.ClusterStatusInstalling),
-				},
-			}
-			Expect(db.Create(&cluster).Error).ToNot(HaveOccurred())
-			mockClusterApi.EXPECT().SetVipsData(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Return(errors.New("Stam"))
-			params := makeStepReply(*clusterId, *hostId, makeResponse("1.2.3.10", "1.2.3.11"))
-			reply := bm.PostStepReply(ctx, params)
-			Expect(reply).Should(BeAssignableToTypeOf(installer.NewPostStepReplyInternalServerError()))
-		})
-	})
-})
-
-var _ = Describe("GetFreeAddresses", func() {
-	var (
-		bm                  *bareMetalInventory
-		cfg                 Config
-		db                  *gorm.DB
-		ctx                 = context.Background()
-		ctrl                *gomock.Controller
-		mockHostApi         *host.MockAPI
-		mockEvents          *events.MockHandler
-		mockSecretValidator *validations.MockPullSecretValidator
-		dbName              = "get_free_addresses"
-	)
-
-	BeforeEach(func() {
-		Expect(envconfig.Process("test", &cfg)).ShouldNot(HaveOccurred())
-		ctrl = gomock.NewController(GinkgoT())
-		db = common.PrepareTestDB(dbName)
-		mockHostApi = host.NewMockAPI(ctrl)
-		mockEvents = events.NewMockHandler(ctrl)
-		mockSecretValidator = validations.NewMockPullSecretValidator(ctrl)
-		bm = NewBareMetalInventory(db, getTestLog(), mockHostApi, nil, cfg, nil, mockEvents, nil, nil, getTestAuthHandler(), nil, nil, mockSecretValidator)
-	})
-
-	AfterEach(func() {
-		ctrl.Finish()
-		common.DeleteTestDB(db, dbName)
-	})
-
-	var makeHost = func(clusterId *strfmt.UUID, freeAddresses, status string) *models.Host {
-		hostId := strToUUID(uuid.New().String())
-		ret := models.Host{
-			ID:            hostId,
-			ClusterID:     *clusterId,
-			FreeAddresses: freeAddresses,
-			Status:        &status,
-		}
-		Expect(db.Create(&ret).Error).ToNot(HaveOccurred())
-		return &ret
-	}
-
-	var makeGetFreeAddressesParams = func(clusterID strfmt.UUID, network string) installer.GetFreeAddressesParams {
-		return installer.GetFreeAddressesParams{
-			ClusterID: clusterID,
-			Network:   network,
-		}
-	}
-
-	It("success", func() {
-		clusterId := strToUUID(uuid.New().String())
-
-		_ = makeHost(clusterId, makeFreeNetworksAddressesStr(makeFreeAddresses("10.0.0.0/16", "10.0.10.1", "10.0.20.0", "10.0.9.250")), models.HostStatusInsufficient)
-		params := makeGetFreeAddressesParams(*clusterId, "10.0.0.0/16")
-		reply := bm.GetFreeAddresses(ctx, params)
-		Expect(reply).Should(BeAssignableToTypeOf(installer.NewGetFreeAddressesOK()))
-		actualReply := reply.(*installer.GetFreeAddressesOK)
-		Expect(len(actualReply.Payload)).To(Equal(3))
-		Expect(actualReply.Payload[0]).To(Equal(strfmt.IPv4("10.0.9.250")))
-		Expect(actualReply.Payload[1]).To(Equal(strfmt.IPv4("10.0.10.1")))
-		Expect(actualReply.Payload[2]).To(Equal(strfmt.IPv4("10.0.20.0")))
-	})
-
-	It("success with limit", func() {
-		clusterId := strToUUID(uuid.New().String())
-
-		_ = makeHost(clusterId, makeFreeNetworksAddressesStr(makeFreeAddresses("10.0.0.0/16", "10.0.10.1", "10.0.20.0", "10.0.9.250")), models.HostStatusInsufficient)
-		params := makeGetFreeAddressesParams(*clusterId, "10.0.0.0/16")
-		params.Limit = swag.Int64(2)
-		reply := bm.GetFreeAddresses(ctx, params)
-		Expect(reply).Should(BeAssignableToTypeOf(installer.NewGetFreeAddressesOK()))
-		actualReply := reply.(*installer.GetFreeAddressesOK)
-		Expect(len(actualReply.Payload)).To(Equal(2))
-		Expect(actualReply.Payload[0]).To(Equal(strfmt.IPv4("10.0.9.250")))
-		Expect(actualReply.Payload[1]).To(Equal(strfmt.IPv4("10.0.10.1")))
-	})
-
-	It("success with limit and prefix", func() {
-		clusterId := strToUUID(uuid.New().String())
-
-		_ = makeHost(clusterId, makeFreeNetworksAddressesStr(makeFreeAddresses("10.0.0.0/16", "10.0.10.1", "10.0.20.0", "10.0.9.250", "10.0.1.0")), models.HostStatusInsufficient)
-		params := makeGetFreeAddressesParams(*clusterId, "10.0.0.0/16")
-		params.Limit = swag.Int64(2)
-		params.Prefix = swag.String("10.0.1")
-		reply := bm.GetFreeAddresses(ctx, params)
-		Expect(reply).Should(BeAssignableToTypeOf(installer.NewGetFreeAddressesOK()))
-		actualReply := reply.(*installer.GetFreeAddressesOK)
-		Expect(len(actualReply.Payload)).To(Equal(2))
-		Expect(actualReply.Payload[0]).To(Equal(strfmt.IPv4("10.0.1.0")))
-		Expect(actualReply.Payload[1]).To(Equal(strfmt.IPv4("10.0.10.1")))
-	})
-
-	It("one disconnected", func() {
-		clusterId := strToUUID(uuid.New().String())
-
-		_ = makeHost(clusterId, makeFreeNetworksAddressesStr(makeFreeAddresses("10.0.0.0/24", "10.0.0.0", "10.0.0.1")), models.HostStatusInsufficient)
-		_ = makeHost(clusterId, makeFreeNetworksAddressesStr(makeFreeAddresses("10.0.0.0/24", "10.0.0.0", "10.0.0.2")), models.HostStatusKnown)
-		_ = makeHost(clusterId, makeFreeNetworksAddressesStr(makeFreeAddresses("10.0.0.0/24")), models.HostStatusDisconnected)
-		params := makeGetFreeAddressesParams(*clusterId, "10.0.0.0/24")
-		reply := bm.GetFreeAddresses(ctx, params)
-		Expect(reply).Should(BeAssignableToTypeOf(installer.NewGetFreeAddressesOK()))
-		actualReply := reply.(*installer.GetFreeAddressesOK)
-		Expect(len(actualReply.Payload)).To(Equal(1))
-		Expect(actualReply.Payload).To(ContainElement(strfmt.IPv4("10.0.0.0")))
-	})
-
-	It("empty result", func() {
-		clusterId := strToUUID(uuid.New().String())
-
-		_ = makeHost(clusterId, makeFreeNetworksAddressesStr(makeFreeAddresses("192.168.0.0/24"),
-			makeFreeAddresses("10.0.0.0/24", "10.0.0.0", "10.0.0.1")), models.HostStatusInsufficient)
-		_ = makeHost(clusterId, makeFreeNetworksAddressesStr(makeFreeAddresses("10.0.0.0/24", "10.0.0.0", "10.0.0.2"),
-			makeFreeAddresses("192.168.0.0/24")), models.HostStatusKnown)
-		_ = makeHost(clusterId, makeFreeNetworksAddressesStr(makeFreeAddresses("10.0.0.0/24", "10.0.0.1", "10.0.0.2")), models.HostStatusInsufficient)
-		params := makeGetFreeAddressesParams(*clusterId, "10.0.0.0/24")
-		reply := bm.GetFreeAddresses(ctx, params)
-		Expect(reply).Should(BeAssignableToTypeOf(installer.NewGetFreeAddressesOK()))
-		actualReply := reply.(*installer.GetFreeAddressesOK)
-		Expect(actualReply.Payload).To(BeEmpty())
-	})
-
-	It("malformed", func() {
-		clusterId := strToUUID(uuid.New().String())
-
-		_ = makeHost(clusterId, makeFreeNetworksAddressesStr(makeFreeAddresses("192.168.0.0/24"),
-			makeFreeAddresses("10.0.0.0/24", "10.0.0.0", "10.0.0.1")), models.HostStatusInsufficient)
-		_ = makeHost(clusterId, makeFreeNetworksAddressesStr(makeFreeAddresses("10.0.0.0/24", "10.0.0.0", "10.0.0.2"),
-			makeFreeAddresses("192.168.0.0/24")), models.HostStatusKnown)
-		_ = makeHost(clusterId, "blah ", models.HostStatusInsufficient)
-		params := makeGetFreeAddressesParams(*clusterId, "10.0.0.0/24")
-		reply := bm.GetFreeAddresses(ctx, params)
-		Expect(reply).Should(BeAssignableToTypeOf(installer.NewGetFreeAddressesOK()))
-		actualReply := reply.(*installer.GetFreeAddressesOK)
-		Expect(len(actualReply.Payload)).To(Equal(1))
-		Expect(actualReply.Payload).To(ContainElement(strfmt.IPv4("10.0.0.0")))
-	})
-
-	It("no matching  hosts", func() {
-		clusterId := strToUUID(uuid.New().String())
-
-		_ = makeHost(clusterId, makeFreeNetworksAddressesStr(makeFreeAddresses("192.168.0.0/24"),
-			makeFreeAddresses("10.0.0.0/24", "10.0.0.0", "10.0.0.1")), models.HostStatusDisconnected)
-		_ = makeHost(clusterId, makeFreeNetworksAddressesStr(makeFreeAddresses("10.0.0.0/24", "10.0.0.0", "10.0.0.2"),
-			makeFreeAddresses("192.168.0.0/24")), models.HostStatusDiscovering)
-		_ = makeHost(clusterId, makeFreeNetworksAddressesStr(makeFreeAddresses("10.0.0.1/24", "10.0.0.0", "10.0.0.2")), models.HostStatusInstalling)
-		params := makeGetFreeAddressesParams(*clusterId, "10.0.0.0/24")
-		verifyApiError(bm.GetFreeAddresses(ctx, params), http.StatusNotFound)
-	})
-})
-
 var _ = Describe("UpdateHostInstallProgress", func() {
 	var (
 		bm                   *bareMetalInventory
