+++ internal/connectivity/mock_connectivity_validator.go
@@ -5,10 +5,9 @@
 package connectivity
 
 import (
-	reflect "reflect"
-
 	gomock "github.com/golang/mock/gomock"
 	models "github.com/openshift/assisted-service/models"
+	reflect "reflect"
 )
 
 // MockValidator is a mock of Validator interface
