+++ pkg/k8sclient/mock_k8sclient.go
@@ -5,11 +5,10 @@
 package k8sclient
 
 import (
-	reflect "reflect"
-
 	gomock "github.com/golang/mock/gomock"
 	v1 "github.com/openshift/api/config/v1"
 	v10 "k8s.io/api/core/v1"
+	reflect "reflect"
 )
 
 // MockK8SClient is a mock of K8SClient interface
